# API_DOCUMENTATION.yaml
openapi: 3.0.3
info:
  title: Task Management & Workflow Orchestration API
  description: |
    Comprehensive Task Management System with Generic Workflow Orchestration.
    
    **Architecture**: Hybrid REST + Event-driven microservices
    
    **Key Features**:
    - Workflow creation and management
    - Team and member management
    - Task assignment and tracking
    - Stage-based orchestration (Process & Escalation workflows)
    - Automatic reassignment when stages change teams
    - SLA configuration with dynamic priority levels
    - Automated task orchestration (workflow selection, SLA configuration, member assignment)
    - Workload-based task assignment
    - Priority rule engine with condition-based evaluation
    
    **All endpoints are routed through API Gateway (Port 5004)**
  version: 3.0.0
  contact:
    name: API Support
    email: support@workflowautomation.enginuo.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:5004/api
    description: Local Development (API Gateway)
  - url: https://api.workflowautomation.enginuo.com/api
    description: Production (API Gateway)
...
paths:
...
  /tasks/{id}:
    get:
      tags: [Tasks]
      summary: Get task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Tasks]
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Tasks]
      summary: Delete task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Task deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
  /tasks/workflow/{workflowId}:
    get:
      tags: [Tasks]
      summary: Get tasks by workflow
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of tasks for workflow
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
  /tasks/stage/{stageId}:
    get:
      tags: [Tasks]
      summary: Get tasks by stage
      parameters:
        - name: stageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of tasks for stage
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
  /tasks/member/{memberId}:
    get:
      tags: [Tasks]
      summary: Get tasks by member
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of tasks for member
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'

  # TaskService (Orchestration)
  /task-service:
    post:
      tags: [TaskService]
      summary: Create task (entry point for orchestration flow)
      description: |
        Creates a new task and initiates the event-driven orchestration flow.
        Returns taskId immediately. Use GET /api/task-service/{id} to retrieve full details after processing.
        
        **Orchestration Flow**:
        1. TaskCreatedEvent published
        2. WorkflowService selects workflow
        3. PriorityRuleEngine assigns priority
        4. SLAManagerService configures SLA
        5. WorkloadService assigns task to best member
        6. WorkflowService starts task in first stage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '202':
          description: Task creation initiated (processing asynchronously)
          content:
            application/json:
              schema:
                type: object
                properties:
                  taskId:
                    type: string
                    format: uuid
                  message:
                    type: string
                    example: "Task creation initiated. Processing asynchronously. Use GET /api/task-service/{taskId} to retrieve full details."
  /task-service/{id}:
    get:
      tags: [TaskService]
      summary: Get task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskServiceTask'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [TaskService]
      summary: Delete task
      description: Deletes task from both TaskService database and WorkflowManagement database
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Task deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
  /task-service/status/{id}:
    put:
      tags: [TaskService]
      summary: Update task status
      description: |
        Updates task status. Accepts string status values or numeric values (0-8) for backward compatibility.
        
        **Valid Status Values**:
        - "Created" (0)
        - "WorkflowSelected" (1)
        - "SLAConfigured" (2)
        - "Assigned" (3)
        - "InProgress" (4)
        - "Completed" (5)
        - "Overdue" (6)
        - "Cancelled" (7)
        - "Escalated" (8)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskStatus'
      responses:
        '200':
          description: Task status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Task status updated to Completed"
        '404':
          $ref: '#/components/responses/NotFound'
  /task-service/complete-stage/{id}:
    post:
      tags: [TaskService]
      summary: Complete current stage and transition to next
      description: |
        Completes the current stage of a task and transitions to the next stage.
        Publishes TaskStageCompletedEvent which triggers stage transition.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Stage completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Stage completed successfully. Task will transition to the next stage."
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          description: Cannot complete stage (e.g., no current stage, already completed)
  /task-service/escalate-stage/{id}:
    post:
      tags: [TaskService]
      summary: Manually escalate task to next stage
      description: |
        Escalates the current stage of a task to the next stage.
        Unlike CompleteStage, this does NOT increment the member's completion count.
        Used when a member cannot handle the task and needs to pass it to the next stage.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Reason for escalation
      responses:
        '200':
          description: Stage escalated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Stage escalated successfully. Task will transition to the next stage."
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          description: Cannot escalate stage (e.g., no current stage, no next stage)
...
components:
  schemas:
...
    Task:
      type: object
      properties:
        taskId:
          type: string
          format: uuid
        taskName:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [Created, WorkflowSelected, SLAConfigured, Assigned, InProgress, Completed, Overdue, Cancelled, Escalated]
        priority:
          type: string
        dueDate:
          type: string
          format: date-time
          nullable: true
        workflowId:
          type: integer
        stageId:
          type: integer
          nullable: true
        stageName:
          type: string
          nullable: true
        assignedToMemberId:
          type: integer
          nullable: true
        assignedToMemberName:
          type: string
          nullable: true
        isOverdue:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TaskUpdate:
      type: object
      required: [taskName, status, priority]
      properties:
        taskName:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        status:
          type: string
        priority:
          type: string
        dueDate:
          type: string
          format: date-time
          nullable: true
        stageId:
          type: integer
          nullable: true
        assignedToMemberId:
          type: integer
          nullable: true
    TaskCreate:
      type: object
      required: [taskName, priority, taskType]
      properties:
        taskName:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        priority:
          type: string
        taskType:
          type: string
        taskData:
          type: object
          nullable: true
          additionalProperties: true
    TaskServiceTask:
      type: object
      properties:
        taskId:
          type: string
          format: uuid
        taskName:
          type: string
        description:
          type: string
          nullable: true
        priority:
          type: string
        taskType:
          type: string
        status:
          type: integer
          enum: [0, 1, 2, 3, 4, 5, 6, 7, 8]
          description: "0=Created, 1=WorkflowSelected, 2=SLAConfigured, 3=Assigned, 4=InProgress, 5=Completed, 6=Overdue, 7=Cancelled, 8=Escalated"
        workflowId:
          type: integer
          nullable: true
        memberId:
          type: integer
          nullable: true
        slaConfigurationId:
          type: string
          format: uuid
          nullable: true
        slaStartTime:
          type: string
          format: date-time
          nullable: true
        slaDeadline:
          type: string
          format: date-time
          nullable: true
        isOverdue:
          type: boolean
        currentStageId:
          type: integer
          nullable: true
        currentStageStartedAt:
          type: string
          format: date-time
          nullable: true
        stageTimeoutAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateTaskStatus:
      type: object
      required: [status]
      properties:
        status:
          oneOf:
            - type: string
              enum: [Created, WorkflowSelected, SLAConfigured, Assigned, InProgress, Completed, Overdue, Cancelled, Escalated]
            - type: integer
              enum: [0, 1, 2, 3, 4, 5, 6, 7, 8]
          description: "String status name or numeric value (0-8) for backward compatibility"
...
    Error:
      type: object
      properties:
        error:
          type: string
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - ApiKey: []
  - BearerAuth: []